mock_sso/requirements.txt
fastapi>=0.90.0,<0.95.0
uvicorn[standard]>=0.20.0,<1.0.0
python-jose>=3.3.0,<4.0.0
python-multipart>=0.0.20

